Static builds for Linux ARM and x64 + Download Windows Portable builds
This is a workflow based on the upstream workflow, release.yml. It is needed because:

- Upstream doesn't publish ARM64 binaries for Linux
- The Windows portable releases are not part of the artifacts attached to the GitHub release

The goal here is to have 4 standalone binaries - Linux x64, Linux ARM, Win x64,
Win ARM. These are cobbled together and attached to a draft release which we can
then publish and use to pull these binaries.